+++ src/components/form/__internal__/form-summary.style.ts
@@ -1,6 +1,5 @@
 import styled, { css } from "styled-components";
 import StyledIcon from "../../icon/icon.style";
-import StyledButton from "../../button/button.style";
 
 type StyledFormSummaryProps = {
   showSummary?: boolean;
@@ -9,12 +8,11 @@ type StyledFormSummaryProps = {
 
 export const StyledFormSummary = styled.div<StyledFormSummaryProps>`
   display: inline-flex;
+  flex-wrap: wrap;
   align-items: center;
-  font-size: 13px;
-  font-weight: 700;
-  margin: -8px;
-  padding: 8px;
-  white-space: nowrap;
+  justify-content: end;
+  font-size: var(--fontSizes100);
+  font-weight: 500;
 
   ${({ fullWidth }) =>
     fullWidth &&
@@ -25,21 +23,27 @@ export const StyledFormSummary = styled.div<StyledFormSummaryProps>`
       justify-content: flex-start;
     `}
 
-  ${({ showSummary }) =>
+  ${({ showSummary, fullWidth }) =>
     showSummary &&
     css`
       background-color: var(--colorsUtilityMajor025);
+      border: solid var(--borderWidth100) var(--colorsActionMinor250);
+      border-radius: var(--borderRadius100);
+      margin: calc(-1 * var(--sizing100)) 0;
+      padding: var(--sizing100);
+      gap: var(--sizing125);
+
+      ${fullWidth &&
+      css`
+        margin: 0 calc(-1 * var(--sizing100));
+      `}
     `}
-  ${StyledButton} {
-    margin-right: 0;
-  }
 `;
 
 export const StyledMessagePrefix = styled.div`
   &:first-of-type {
-    margin-left: 4px;
+    margin-left: var(--sizing100);
   }
-  margin-right: 4px;
 `;
 
 export type StyledInternalSummaryProps = {
@@ -49,10 +53,7 @@ export type StyledInternalSummaryProps = {
 export const StyledInternalSummary = styled.div<StyledInternalSummaryProps>`
   display: flex;
   align-items: center;
-  margin-right: 8px;
-  &:last-of-type {
-    margin-right: 16px;
-  }
+  gap: var(--sizing100);
   ${({ type }) =>
     type === "warning" &&
     css`
@@ -65,7 +66,6 @@ export const StyledInternalSummary = styled.div<StyledInternalSummaryProps>`
     `}
 
   ${StyledIcon} {
-    margin-right: 4px;
     ${({ type }) =>
       type === "warning" &&
       css`