+++ src/components/select/multi-select/multi-select.spec.tsx
@@ -1231,7 +1231,7 @@ describe("MultiSelect", () => {
         {
           content: '"*"',
           color: "var(--colorsSemanticNegative500)",
-          fontWeight: "var(--fontWeights700)",
+          fontWeight: "var(--fontWeights500)",
           marginLeft: "var(--spacing050)",
         },
         wrapper.find(StyledLabel),