+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+### [142.13.1](https://github.com/Sage/carbon/compare/v142.13.0...v142.13.1) (2024-10-03)
+
+
+### Bug Fixes
+
+* **form:** allow form-footer content to be responsive ([c7d55e1](https://github.com/Sage/carbon/commit/c7d55e14a44245609a5bd94df8993eb5a6f4b2ba)), closes [#6658](https://github.com/Sage/carbon/issues/6658)
+
+## [142.13.0](https://github.com/Sage/carbon/compare/v142.12.0...v142.13.0) (2024-10-02)
+
+
+### Features
+
+* **fonts:** updated font weights for ds base token alignment ([58e89d8](https://github.com/Sage/carbon/commit/58e89d8255d2ff288adef2ddc03cb2679dabe481))
+
 ## [142.12.0](https://github.com/Sage/carbon/compare/v142.11.8...v142.12.0) (2024-09-26)